+++ web/packages/accounts/pages/two-factor/recover.tsx
@@ -9,9 +9,9 @@ import {
     removeTwoFactor,
     type TwoFactorType,
 } from "@/accounts/services/user";
-import type { AccountsContextT } from "@/accounts/types/context";
 import { LinkButton } from "@/base/components/LinkButton";
 import type { MiniDialogAttributes } from "@/base/components/MiniDialog";
+import type { AccountsContextT } from "@/base/context";
 import { sharedCryptoWorker } from "@/base/crypto";
 import type { B64EncryptionResult } from "@/base/crypto/libsodium";
 import log from "@/base/log";
